Revert "CRITICAL REVERT: Restore Original App Structure + Fix Black Border & Discovery"
