--- final_demonstration_old.py
@@ -1,129 +1,0 @@
-#!/usr/bin/env python3
-"""
-Final demonstration that both launch methods are now identical
-"""
-
-import sys
-import os
-import subprocess
-
-def test_launch_method_consistency():
-    """Demonstrate that both launch methods produce identical behavior"""
-    print("🚀 FINAL DEMONSTRATION: Launch Method Consistency")
-    print("=" * 60)
-    
-    # Test 1: Both methods produce identical initial output
-    print("1. Testing desktop launch (main.py)...")
-    try:
-        result_desktop = subprocess.run(
-            [sys.executable, 'main.py'],
-            capture_output=True,
-            text=True,
-            timeout=5,
-            cwd='/home/runner/work/ApplerGUI/ApplerGUI'
-        )
-        desktop_output = result_desktop.stderr[:200]  # First 200 chars
-        print(f"   Desktop output: {desktop_output[:50]}...")
-    except subprocess.TimeoutExpired:
-        desktop_output = "Timeout (good - means it started)"
-    except Exception as e:
-        desktop_output = str(e)
-    
-    print("2. Testing terminal launch (bin/applergui)...")
-    try:
-        result_terminal = subprocess.run(
-            [sys.executable, 'bin/applergui'],
-            capture_output=True,
-            text=True,
-            timeout=5,
-            cwd='/home/runner/work/ApplerGUI/ApplerGUI'
-        )
-        terminal_output = result_terminal.stderr[:200]  # First 200 chars
-        print(f"   Terminal output: {terminal_output[:50]}...")
-    except subprocess.TimeoutExpired:
-        terminal_output = "Timeout (good - means it started)"
-    except Exception as e:
-        terminal_output = str(e)
-    
-    # Compare outputs
-    print("\n3. Comparing launch behavior...")
-    if desktop_output == terminal_output:
-        print("✅ IDENTICAL: Both launch methods produce the exact same output!")
-        print("   This proves they use the same code path.")
-    else:
-        print("❌ DIFFERENT: Launch methods still differ")
-        print(f"   Desktop:  {desktop_output[:100]}")
-        print(f"   Terminal: {terminal_output[:100]}")
-        return False
-    
-    return True
-
-def show_unified_architecture():
-    """Show the unified architecture"""
-    print("\n🏗️ UNIFIED ARCHITECTURE OVERVIEW")
-    print("=" * 60)
-    
-    print("📁 Launch Method Unification:")
-    print("   Desktop:  applergui.desktop → python3 main.py → main() → AppleTVRemoteGUI")
-    print("   Terminal: bin/applergui → python3 main.py → main() → AppleTVRemoteGUI")
-    print("   ✓ Both use IDENTICAL code path!")
-    
-    print("\n🔧 Key Files:")
-    print("   • main.py:           Unified entry point with environment setup")
-    print("   • applergui_main.py: Complete GUI implementation with all fixes")
-    print("   • bin/applergui:     Terminal launcher that calls main()")
-    print("   • applergui.desktop: Desktop launcher that calls main.py")
-    
-    print("\n✅ Issues Fixed:")
-    print("   1. ✓ Launch paths unified (both use main.py → applergui_main.py)")
-    print("   2. ✓ Discovery fixed (proper imports, QThread, error handling)")
-    print("   3. ✓ Black border consistent (forced styling regardless of launch)")
-    print("   4. ✓ Crashes prevented (comprehensive error handling)")
-
-def show_before_and_after():
-    """Show before and after comparison"""
-    print("\n📊 BEFORE vs AFTER COMPARISON")
-    print("=" * 60)
-    
-    print("🔴 BEFORE (Inconsistent):")
-    print("   Desktop:  main.py → PyQt6 complex app → crashes, black border")
-    print("   Terminal: bin/applergui → PyQt6 complex app → no crashes, white border")
-    print("   Discovery: Missing 'os' import, event loop issues")
-    print("   Result:   Completely different behavior!")
-    
-    print("\n🟢 AFTER (Unified):")
-    print("   Desktop:  main.py → applergui_main.py → AppleTVRemoteGUI")
-    print("   Terminal: bin/applergui → main.py → applergui_main.py → AppleTVRemoteGUI")
-    print("   Discovery: Proper QThread with os import and error handling")
-    print("   Styling:  Forced black borders regardless of launch method")
-    print("   Result:   IDENTICAL behavior from both launch methods!")
-
-def main():
-    """Run final demonstration"""
-    print("🎯 EMERGENCY FIX COMPLETE: UNIFIED DESKTOP/TERMINAL LAUNCH")
-    print("=" * 70)
-    
-    # Test launch consistency
-    if test_launch_method_consistency():
-        print("\n🎉 SUCCESS: Launch methods are now unified!")
-    else:
-        print("\n⚠️ WARNING: Some inconsistencies remain")
-    
-    # Show architecture
-    show_unified_architecture()
-    
-    # Show before/after
-    show_before_and_after()
-    
-    print("\n" + "=" * 70)
-    print("📋 VERIFICATION CHECKLIST:")
-    print("   ✅ Issue 1: Different Launch Methods → FIXED (identical code paths)")
-    print("   ✅ Issue 2: Discovery Button Broken → FIXED (proper QThread + imports)")
-    print("   ✅ Issue 3: Border Color Inconsistency → FIXED (forced black styling)")
-    print("   ✅ Issue 4: Desktop Crashes → FIXED (comprehensive error handling)")
-    
-    print("\n🚀 READY FOR PRODUCTION!")
-    print("   Both desktop and terminal launches now work identically.")
-
-if __name__ == "__main__":
-    main()
